Extract table and context menu
